add comment service
fix unit test

+++ server/speciality/src/test/java/com/df/spec/locality/dao/MongoShopDaoTest.java
@@ -23,18 +23,16 @@ public class MongoShopDaoTest extends SpecialityBaseTest {
 		return region;
 	}
 
-	protected void removeRegion(Region region) {
-		regionDao.deleteRegion(region);
+	protected void removeRegion() {
+		regionDao.deleteRegion(new Region("上海市", "上海市", "黄浦区"));
 	}
 
 	@Test
 	public void testAddShop() {
-		Region region = this.createRegion();
-		Shop shop = new Shop("城隍庙第一百货", "黄浦区安仁街218号");
+		Shop shop = new Shop("城隍庙第一百货测试", "黄浦区安仁街218号");
 		try {
-
+			Region region = this.createRegion();
 			shopDao.addShop(shop, region);
-			shopDao.deleteShop(shop.getCode());
 			Shop shop2 = new Shop(shop.getName(), shop.getAddress());
 			try {
 				shopDao.addShop(shop2, region);
@@ -43,7 +41,7 @@ public class MongoShopDaoTest extends SpecialityBaseTest {
 			}
 			TestCase.fail();
 		} finally {
-			this.removeRegion(region);
+			this.removeRegion();
 			if (shop.getCode() != null) {
 				shopDao.deleteShop(shop.getCode());
 			}
@@ -52,14 +50,16 @@ public class MongoShopDaoTest extends SpecialityBaseTest {
 
 	@Test
 	public void testGetShopByCode() {
-		Region region = this.createRegion();
-		Shop shop = new Shop("城隍庙第一百货", "黄浦区安仁街218号");
+		Shop shop = new Shop("城隍庙第一百货测试", "黄浦区安仁街218号");
 		try {
+			Region region = this.createRegion();
 			shopDao.addShop(shop, region);
 			TestCase.assertNotNull(shopDao.getShopByCode(shop.getCode()));
 		} finally {
-			this.removeRegion(region);
-			shopDao.deleteShop(shop.getCode());
+			this.removeRegion();
+			if (shop.getCode() != null) {
+				shopDao.deleteShop(shop.getCode());
+			}
 		}
 	}
 }
